clean up city section list query renderer
