+++ src/lib/Containers/index.tsx
@@ -1,5 +1,4 @@
 import { ArtworkContainer as Artwork } from "lib/Scenes/Artwork"
-import CitySectionList from "lib/Scenes/City/CitySectionList"
 import { PartnerContainer as Partner } from "lib/Scenes/Partner"
 import Fair from "../Scenes/Fair"
 import { Home } from "../Scenes/Home/Home"
@@ -8,7 +7,6 @@ import Show from "../Scenes/Show"
 
 export default {
   Artwork,
-  CitySectionList,
   Fair,
   Home,
   MyProfile,
